+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -1923,6 +1923,9 @@
     "name": "init_zoneless_scheduling"
   },
   {
+    "name": "init_zoneless_scheduling_impl"
+  },
+  {
     "name": "initializeDirectives"
   },
   {
